fix(viewModel) : 웹 크롤링 동작코드는 data영역으로 이동
